+++ drivers/block/drbd/drbd_nl.c
@@ -3361,7 +3361,7 @@ put_result:
 	err = resource_statistics_to_skb(skb, &resource_statistics, !capable(CAP_SYS_ADMIN));
 	if (err)
 		goto out;
-	cb->args[0] = (long)resource;
+	cb->args[0] = (intptr_t)resource;
 	genlmsg_end(skb, dh);
 	err = 0;
 
@@ -3444,7 +3444,7 @@ int drbd_adm_dump_devices(struct sk_buff *skb, struct netlink_callback *cb)
 				rcu_read_lock();
 				goto put_result;
 			}
-			cb->args[0] = (long)resource;
+			cb->args[0] = (intptr_t)resource;
 		}
 	}
 
@@ -3533,7 +3533,7 @@ int drbd_adm_dump_connections(struct sk_buff *skb, struct netlink_callback *cb)
 			resource = drbd_find_resource(nla_data(resource_filter));
 			if (!resource)
 				goto put_result;
-			cb->args[0] = (long)resource;
+			cb->args[0] = (intptr_t)resource;
 			cb->args[1] = SINGLE_RESOURCE;
 		}
 	}
@@ -3542,7 +3542,7 @@ int drbd_adm_dump_connections(struct sk_buff *skb, struct netlink_callback *cb)
 			goto out;
 		resource = list_first_entry(&drbd_resources, struct drbd_resource, resources);
 		kref_get(&resource->kref);
-		cb->args[0] = (long)resource;
+		cb->args[0] = (intptr_t)resource;
 		cb->args[1] = ITERATE_RESOURCES;
 	}
 
@@ -3583,7 +3583,7 @@ found_resource:
 		kref_put(&resource->kref, drbd_destroy_resource);
 		resource = next_resource;
 		kref_get(&resource->kref);
-		cb->args[0] = (long)resource;
+		cb->args[0] = (intptr_t)resource;
 		cb->args[2] = 0;
 		goto next_resource;
 	}
@@ -3618,7 +3618,7 @@ put_result:
 		err = connection_statistics_to_skb(skb, &connection_statistics, !capable(CAP_SYS_ADMIN));
 		if (err)
 			goto out;
-		cb->args[2] = (long)connection;
+		cb->args[2] = (intptr_t)connection;
 	}
 	genlmsg_end(skb, dh);
 	err = 0;
@@ -3697,7 +3697,7 @@ int drbd_adm_dump_peer_devices(struct sk_buff *skb, struct netlink_callback *cb)
 				goto put_result;
 			}
 		}
-		cb->args[0] = (long)resource;
+		cb->args[0] = (intptr_t)resource;
 	}
 
 	rcu_read_lock();
@@ -3759,7 +3759,7 @@ put_result:
 		if (err)
 			goto out;
 		cb->args[1] = minor;
-		cb->args[2] = (long)peer_device;
+		cb->args[2] = (intptr_t)peer_device;
 	}
 	genlmsg_end(skb, dh);
 	err = 0;
@@ -4041,7 +4041,7 @@ done:
 out:
 	rcu_read_unlock();
 	/* where to start the next iteration */
-	cb->args[0] = (long)pos;
+	cb->args[0] = (intptr_t)pos;
 	cb->args[1] = (pos == resource) ? volume + 1 : 0;
 
 	/* No more resources/volumes/minors found results in an empty skb.
@@ -4099,9 +4099,9 @@ int drbd_adm_get_status_all(struct sk_buff *skb, struct netlink_callback *cb)
 
 	/* prime iterators, and set "filter" mode mark:
 	 * only dump this connection. */
-	cb->args[0] = (long)resource;
+	cb->args[0] = (intptr_t)resource;
 	/* cb->args[1] = 0; passed in this way. */
-	cb->args[2] = (long)resource;
+	cb->args[2] = (intptr_t)resource;
 
 dump:
 	return get_one_status(skb, cb);
@@ -4976,7 +4976,7 @@ next:
 		struct drbd_state_change *next_state_change =
 			list_entry(state_change->list.next,
 				   struct drbd_state_change, list);
-		cb->args[0] = (long)next_state_change;
+		cb->args[0] = (intptr_t)next_state_change;
 		cb->args[3] = notifications_for_state_change(next_state_change);
 		cb->args[4] = 0;
 	}
@@ -5027,7 +5027,7 @@ int drbd_adm_get_initial_state(struct sk_buff *skb, struct netlink_callback *cb)
 	if (!list_empty(&head)) {
 		struct drbd_state_change *state_change =
 			list_entry(head.next, struct drbd_state_change, list);
-		cb->args[0] = (long)state_change;
+		cb->args[0] = (intptr_t)state_change;
 		cb->args[3] = notifications_for_state_change(state_change);
 		list_del(&head);  /* detach list from head */
 	}
