+++ drivers/gpu/drm/xe/tests/xe_gt_sriov_pf_config_kunit.c
@@ -33,7 +33,7 @@ static void pf_set_usable_vram(struct xe_device *xe, u64 usable)
 
 static const void *num_vfs_gen_param(struct kunit *test, const void *prev, char *desc)
 {
-	unsigned long next = 1 + (unsigned long)prev;
+	unsigned long next = 1 + (uintptr_t)prev;
 
 	if (next > TEST_MAX_VFS)
 		return NULL;
