+++ include/linux/skbuff.h
@@ -1218,7 +1218,7 @@ static inline void skb_dst_set(struct sk_buff *skb, struct dst_entry *dst)
 {
 	skb_dst_check_unset(skb);
 	skb->slow_gro |= !!dst;
-	skb->_skb_refdst = (unsigned long)dst;
+	skb->_skb_refdst = (uintptr_t)dst;
 }
 
 /**
@@ -1236,7 +1236,7 @@ static inline void skb_dst_set_noref(struct sk_buff *skb, struct dst_entry *dst)
 	skb_dst_check_unset(skb);
 	WARN_ON(!rcu_read_lock_held() && !rcu_read_lock_bh_held());
 	skb->slow_gro |= !!dst;
-	skb->_skb_refdst = (unsigned long)dst | SKB_DST_NOREF;
+	skb->_skb_refdst = (uintptr_t)dst | SKB_DST_NOREF;
 }
 
 /**
