+++ kernel/sched/debug.c
@@ -338,7 +338,7 @@ static ssize_t sched_server_write_common(struct file *filp, const char __user *u
 					 size_t cnt, loff_t *ppos, enum dl_param param,
 					 void *server)
 {
-	long cpu = (long) ((struct seq_file *) filp->private_data)->private;
+	intptr_t cpu = (intptr_t) ((struct seq_file *) filp->private_data)->private;
 	struct sched_dl_entity *dl_se = (struct sched_dl_entity *)server;
 	u64 old_runtime, runtime, period;
 	struct rq *rq = cpu_rq(cpu);
@@ -417,7 +417,7 @@ static ssize_t
 sched_fair_server_runtime_write(struct file *filp, const char __user *ubuf,
 				size_t cnt, loff_t *ppos)
 {
-	long cpu = (long) ((struct seq_file *) filp->private_data)->private;
+	intptr_t cpu = (intptr_t) ((struct seq_file *) filp->private_data)->private;
 	struct rq *rq = cpu_rq(cpu);
 
 	return sched_server_write_common(filp, ubuf, cnt, ppos, DL_RUNTIME,
@@ -426,7 +426,7 @@ sched_fair_server_runtime_write(struct file *filp, const char __user *ubuf,
 
 static int sched_fair_server_runtime_show(struct seq_file *m, void *v)
 {
-	unsigned long cpu = (unsigned long) m->private;
+	uintptr_t cpu = (uintptr_t) m->private;
 	struct rq *rq = cpu_rq(cpu);
 
 	return sched_server_show_common(m, v, DL_RUNTIME, &rq->fair_server);
@@ -483,7 +483,7 @@ static ssize_t
 sched_fair_server_period_write(struct file *filp, const char __user *ubuf,
 			       size_t cnt, loff_t *ppos)
 {
-	long cpu = (long) ((struct seq_file *) filp->private_data)->private;
+	intptr_t cpu = (intptr_t) ((struct seq_file *) filp->private_data)->private;
 	struct rq *rq = cpu_rq(cpu);
 
 	return sched_server_write_common(filp, ubuf, cnt, ppos, DL_PERIOD,
@@ -492,7 +492,7 @@ sched_fair_server_period_write(struct file *filp, const char __user *ubuf,
 
 static int sched_fair_server_period_show(struct seq_file *m, void *v)
 {
-	unsigned long cpu = (unsigned long) m->private;
+	uintptr_t cpu = (uintptr_t) m->private;
 	struct rq *rq = cpu_rq(cpu);
 
 	return sched_server_show_common(m, v, DL_PERIOD, &rq->fair_server);
